+++ src/pages/TopicInfo.tsx
@@ -1,4 +1,4 @@
-import { Button, Card } from "antd";
+import { Button, Card, Modal } from "antd";
 import React from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import {
@@ -13,12 +13,21 @@ function TopicInfo() {
   const [deleteTopic] = useDeleteTopicMutation();
 
   const handleDeleteClick = async () => {
-    try {
-      await deleteTopic({ title: title }).unwrap();
-      navigate(`/topic/${chapter}`);
-    } catch (error) {
-      console.error(error);
-    }
+    Modal.confirm({
+      title: "주의",
+      content: "정말 이 항목을 삭제하시겠습니까?",
+      okText: "예",
+      okType: "danger",
+      cancelText: "아니오",
+      onOk: async () => {
+        try {
+          await deleteTopic({ title: title }).unwrap();
+          navigate(`/topic/${chapter}`);
+        } catch (error) {
+          console.error(error);
+        }
+      },
+    });
   };
 
   const handleUpdateClick = () => {
@@ -43,8 +52,26 @@ function TopicInfo() {
       <p>분류: {topic?.category}</p>
       <p>시작 년도: {topic?.startDate}</p>
       <p>종료 년도: {topic?.endDate}</p>
-      <p>상세설명: {topic?.detail}</p>
       <p>키워드: {topic?.keywordList}</p>
+      <div
+        style={{
+          border: "1px solid rgba(5, 5, 5, 0.06)",
+          borderRadius: 12,
+          padding: 12,
+        }}
+      >
+        <div
+          style={{
+            borderBottom: "1px solid rgba(5, 5, 5, 0.06)",
+            paddingBottom: 12,
+          }}
+        >
+          상세내용
+        </div>
+        {topic?.detail.split("\n").map((line, index) => {
+          return <p key={index}>{line}</p>;
+        })}
+      </div>
     </Card>
   );
 }
